debt - remove AMD support of VSCode for web
