+++ .vscode/settings.json
@@ -107,6 +107,7 @@
 		"@xterm/headless",
 		"node-pty",
 		"vscode-notebook-renderer",
+		"src/vs/workbench/workbench.web.main.ts",
 		"src/vs/workbench/workbench.web.main.internal.ts"
 	],
