+++ src/vs/nls.ts
@@ -3,10 +3,13 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-// eslint-disable-next-line local/code-import-patterns
-import { getNLSLanguage, getNLSMessages } from './nls.messages.js';
-// eslint-disable-next-line local/code-import-patterns
-export { getNLSLanguage, getNLSMessages } from './nls.messages.js';
+export function getNLSMessages(): string[] {
+	return globalThis._VSCODE_NLS_MESSAGES;
+}
+
+export function getNLSLanguage(): string | undefined {
+	return globalThis._VSCODE_NLS_LANGUAGE;
+}
 
 declare const document: { location?: { hash?: string } } | undefined;
 const isPseudo = getNLSLanguage() === 'pseudo' || (typeof document !== 'undefined' && document.location && typeof document.location.hash === 'string' && document.location.hash.indexOf('pseudo=true') >= 0);
